Removed plugin folders, plugin manager functioning
Installable plugins from github as listed in the text file on
FLIKA_plugins. It's kinda annoying to always have to commit every change
to the text file but oh well.  Other small change to how script editor
and plugin manager are saved under global vars. Could possibly be added
to m.dialogs for simplicity. Menu updated on plugin install. Not tested
on python 2